+++ crates/bcr-ebill-core/src/blockchain/bill/block.rs
@@ -368,7 +368,7 @@ pub struct BillRequestRecourseBlockData {
     pub signing_address: PostalAddress, // address of the endorser
 }
 
-#[derive(BorshSerialize, BorshDeserialize, Debug, Clone, PartialEq)]
+#[derive(BorshSerialize, BorshDeserialize, Debug, Clone, PartialEq, Eq)]
 pub enum BillRecourseReasonBlockData {
     Accept,
     Pay,
@@ -398,6 +398,7 @@ pub struct BillRecourseBlockData {
     pub recoursee: BillIdentityBlockData,
     pub sum: u64,
     pub currency: String,
+    pub recourse_reason: BillRecourseReasonBlockData,
     pub signatory: Option<BillSignatoryBlockData>,
     pub signing_timestamp: u64,
     pub signing_address: PostalAddress, // address of the endorser
@@ -1342,6 +1343,12 @@ impl BillBlock {
             }
             Recourse => {
                 let data: BillRecourseBlockData = self.get_decrypted_block_bytes(bill_keys)?;
+                let reason = match data.recourse_reason {
+                    BillRecourseReasonBlockData::Pay => {
+                        RecourseReason::Pay(data.sum, data.currency.clone())
+                    }
+                    BillRecourseReasonBlockData::Accept => RecourseReason::Accept,
+                };
                 data.validate()?;
                 (
                     data.recourser.node_id,
@@ -1350,6 +1357,7 @@ impl BillBlock {
                         data.recoursee.into(),
                         data.sum,
                         data.currency,
+                        reason,
                     )),
                 )
             }
@@ -1402,9 +1410,9 @@ mod tests {
     use crate::{
         blockchain::bill::tests::get_baseline_identity,
         tests::tests::{
-            TEST_BILL_ID, TEST_PRIVATE_KEY_SECP, TEST_PUB_KEY_SECP, empty_bitcredit_bill,
-            empty_identity_public_data, get_bill_keys, identity_public_data_only_node_id,
-            valid_address,
+            TEST_BILL_ID, TEST_PRIVATE_KEY_SECP, TEST_PUB_KEY_SECP, VALID_PAYMENT_ADDRESS_TESTNET,
+            empty_bitcredit_bill, empty_identity_public_data, get_bill_keys,
+            identity_public_data_only_node_id, invalid_address, valid_address,
         },
     };
     use rstest::rstest;
@@ -1648,7 +1656,7 @@ mod tests {
                 seller: seller.clone().into(),
                 sum: 5000,
                 currency: "sat".to_string(),
-                payment_address: "tb1qteyk7pfvvql2r2zrsu4h4xpvju0nz7ykvguyk0".to_string(),
+                payment_address: VALID_PAYMENT_ADDRESS_TESTNET.to_string(),
                 signatory: None,
                 signing_timestamp: 1731593928,
                 signing_address: seller.postal_address,
@@ -1681,7 +1689,7 @@ mod tests {
                 seller: seller.clone().into(),
                 sum: 5000,
                 currency: "sat".to_string(),
-                payment_address: "tb1qteyk7pfvvql2r2zrsu4h4xpvju0nz7ykvguyk0".to_string(),
+                payment_address: VALID_PAYMENT_ADDRESS_TESTNET.to_string(),
                 signatory: Some(BillSignatoryBlockData {
                     node_id: buyer.node_id.clone(),
                     name: buyer.name.clone(),
@@ -1840,6 +1848,7 @@ mod tests {
                 recoursee: recoursee.clone().into(),
                 sum: 15000,
                 currency: "sat".to_string(),
+                recourse_reason: BillRecourseReasonBlockData::Pay,
                 signatory: None,
                 signing_timestamp: 1731593928,
                 signing_address: recourser.postal_address,
@@ -2013,7 +2022,7 @@ mod tests {
                 buyer: other_party.clone().into(),
                 sum: 5000,
                 currency: "sat".to_string(),
-                payment_address: "tb1qteyk7pfvvql2r2zrsu4h4xpvju0nz7ykvguyk0".to_string(),
+                payment_address: VALID_PAYMENT_ADDRESS_TESTNET.to_string(),
                 signatory: None,
                 signing_timestamp: 1731593928,
                 signing_address: signer.postal_address.clone(),
@@ -2039,7 +2048,7 @@ mod tests {
                 buyer: other_party.clone().into(),
                 sum: 5000,
                 currency: "sat".to_string(),
-                payment_address: "tb1qteyk7pfvvql2r2zrsu4h4xpvju0nz7ykvguyk0".to_string(),
+                payment_address: VALID_PAYMENT_ADDRESS_TESTNET.to_string(),
                 signatory: None,
                 signing_timestamp: 1731593928,
                 signing_address: signer.postal_address.clone(),
@@ -2180,6 +2189,7 @@ mod tests {
                 recoursee: other_party.clone().into(),
                 sum: 15000,
                 currency: "sat".to_string(),
+                recourse_reason: BillRecourseReasonBlockData::Pay,
                 signatory: None,
                 signing_timestamp: 1731593928,
                 signing_address: signer.postal_address.clone(),
@@ -2378,7 +2388,7 @@ mod tests {
                 buyer: other_party.clone().into(),
                 sum: 5000,
                 currency: "sat".to_string(),
-                payment_address: "tb1qteyk7pfvvql2r2zrsu4h4xpvju0nz7ykvguyk0".to_string(),
+                payment_address: VALID_PAYMENT_ADDRESS_TESTNET.to_string(),
                 signatory: Some(BillSignatoryBlockData {
                     node_id: identity_keys.get_public_key(),
                     name: "signatory name".to_string(),
@@ -2407,7 +2417,7 @@ mod tests {
                 buyer: other_party.clone().into(),
                 sum: 5000,
                 currency: "sat".to_string(),
-                payment_address: "tb1qteyk7pfvvql2r2zrsu4h4xpvju0nz7ykvguyk0".to_string(),
+                payment_address: VALID_PAYMENT_ADDRESS_TESTNET.to_string(),
                 signatory: Some(BillSignatoryBlockData {
                     node_id: identity_keys.get_public_key(),
                     name: "signatory name".to_string(),
@@ -2566,6 +2576,7 @@ mod tests {
                 recoursee: other_party.clone().into(),
                 sum: 15000,
                 currency: "sat".to_string(),
+                recourse_reason: BillRecourseReasonBlockData::Pay,
                 signatory: Some(BillSignatoryBlockData {
                     node_id: identity_keys.get_public_key(),
                     name: "signatory name".to_string(),
@@ -2705,35 +2716,416 @@ mod tests {
         }
     }
 
+    fn invalid_bill_identity_block_data() -> BillIdentityBlockData {
+        BillIdentityBlockData {
+            t: ContactType::Person,
+            node_id: TEST_PUB_KEY_SECP.into(),
+            name: "".into(),
+            postal_address: invalid_address(),
+        }
+    }
+
     #[test]
     fn test_valid_bill_identity_block_data() {
         assert_eq!(valid_bill_identity_block_data().validate(), Ok(()));
     }
 
     #[rstest]
-    #[case::invalid_node_id(
-        BillIdentityBlockData { node_id: "invalidkey".into(), ..valid_bill_identity_block_data() },
-        ValidationError::InvalidSecp256k1Key("invalidkey".into())
-    )]
-    #[case::empty_name(
-        BillIdentityBlockData { name: "".into(), ..valid_bill_identity_block_data() },
-        ValidationError::FieldEmpty(Field::Name)
-    )]
-    #[case::blank_name(
-        BillIdentityBlockData { name: "   ".into(), ..valid_bill_identity_block_data() },
-        ValidationError::FieldEmpty(Field::Name)
-    )]
-    #[case::invalid_address(
-        BillIdentityBlockData {
-            postal_address: PostalAddress { country: "".into(), ..valid_address() },
-            ..valid_bill_identity_block_data()
-        },
-        ValidationError::FieldEmpty(Field::Country)
-    )]
-    fn test_invalid_identity_data(
+    #[case::invalid_node_id( BillIdentityBlockData { node_id: "invalidkey".into(), ..valid_bill_identity_block_data() }, ValidationError::InvalidSecp256k1Key("invalidkey".into()))]
+    #[case::empty_name( BillIdentityBlockData { name: "".into(), ..valid_bill_identity_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::blank_name( BillIdentityBlockData { name: "   ".into(), ..valid_bill_identity_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_address( BillIdentityBlockData { postal_address: invalid_address(), ..valid_bill_identity_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    fn test_invalid_bill_identity_block_data(
         #[case] identity: BillIdentityBlockData,
         #[case] expected_error: ValidationError,
     ) {
         assert_eq!(identity.validate(), Err(expected_error));
+    }
+
+    fn valid_bill_signatory_block_data() -> BillSignatoryBlockData {
+        BillSignatoryBlockData {
+            node_id: TEST_PUB_KEY_SECP.into(),
+            name: "Johanna Smith".into(),
+        }
+    }
+
+    fn invalid_bill_signatory_block_data() -> BillSignatoryBlockData {
+        BillSignatoryBlockData {
+            node_id: TEST_PUB_KEY_SECP.into(),
+            name: "".into(),
+        }
+    }
+
+    #[test]
+    fn test_valid_bill_signatory_block_data() {
+        assert_eq!(valid_bill_signatory_block_data().validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_node_id( BillSignatoryBlockData { node_id: "invalidkey".into(), ..valid_bill_signatory_block_data() }, ValidationError::InvalidSecp256k1Key("invalidkey".into()))]
+    #[case::empty_name( BillSignatoryBlockData { name: "".into(), ..valid_bill_signatory_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::blank_name( BillSignatoryBlockData { name: "   ".into(), ..valid_bill_signatory_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_bill_signatory_block_data(
+        #[case] signatory: BillSignatoryBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(signatory.validate(), Err(expected_error));
+    }
+
+    fn valid_bill_issue_block_data() -> BillIssueBlockData {
+        BillIssueBlockData {
+            id: TEST_BILL_ID.into(),
+            country_of_issuing: "AT".into(),
+            city_of_issuing: "Vienna".into(),
+            drawee: valid_bill_identity_block_data(),
+            drawer: valid_bill_identity_block_data(),
+            payee: valid_bill_identity_block_data(),
+            currency: "sat".into(),
+            sum: 500,
+            maturity_date: "2025-11-12".into(),
+            issue_date: "2025-08-12".into(),
+            country_of_payment: "FR".into(),
+            city_of_payment: "Paris".into(),
+            language: "de".into(),
+            files: vec![],
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_bill_issue_block_data() {
+        let bill = valid_bill_issue_block_data();
+        assert_eq!(bill.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::empty_id(BillIssueBlockData { id: "".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Id))]
+    #[case::blank_id(BillIssueBlockData { id: "  ".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Id))]
+    #[case::invalid_maturity_data(BillIssueBlockData { maturity_date: "1234-sdfds".into(), ..valid_bill_issue_block_data() }, ValidationError::InvalidDate)]
+    #[case::invalid_issue_date(BillIssueBlockData { issue_date: "2019-fsds-sdf".into(), ..valid_bill_issue_block_data() }, ValidationError::InvalidDate)]
+    #[case::invalid_sum(BillIssueBlockData { sum: 0, ..valid_bill_issue_block_data() }, ValidationError::InvalidSum)]
+    #[case::invalid_currency(BillIssueBlockData { currency: "invalidcurrency".into(), ..valid_bill_issue_block_data() }, ValidationError::InvalidCurrency)]
+    #[case::empty_country_of_issuing(BillIssueBlockData { country_of_issuing: "".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::CountryOfIssuing))]
+    #[case::blank_country_of_issuing(BillIssueBlockData { country_of_issuing: "  ".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::CountryOfIssuing))]
+    #[case::empty_city_of_issuing(BillIssueBlockData { city_of_issuing: "".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::CityOfIssuing))]
+    #[case::blank_city_of_issuing(BillIssueBlockData { city_of_issuing: "  ".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::CityOfIssuing))]
+    #[case::empty_country_of_payment(BillIssueBlockData { country_of_payment: "".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::CountryOfPayment))]
+    #[case::blank_country_of_payment(BillIssueBlockData { country_of_payment: " ".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::CountryOfPayment))]
+    #[case::empty_city_of_payment(BillIssueBlockData { city_of_payment: "".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::CityOfPayment))]
+    #[case::blank_city_of_payment(BillIssueBlockData { city_of_payment: " ".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::CityOfPayment))]
+    #[case::empty_language(BillIssueBlockData { language: "".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Language))]
+    #[case::blank_language(BillIssueBlockData { language: "   ".into(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Language))]
+    #[case::invalid_signatory(BillIssueBlockData { drawee: invalid_bill_identity_block_data(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_drawee(BillIssueBlockData { drawer: invalid_bill_identity_block_data(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_drawer(BillIssueBlockData { payee: invalid_bill_identity_block_data(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_payee(BillIssueBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_signing_address(BillIssueBlockData { signing_address: invalid_address(), ..valid_bill_issue_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    fn test_invalid_bill_issue_block_data(
+        #[case] bill: BillIssueBlockData,
+        #[case] expected: ValidationError,
+    ) {
+        assert_eq!(bill.validate(), Err(expected));
+    }
+
+    fn valid_req_to_accept_block_data() -> BillRequestToAcceptBlockData {
+        BillRequestToAcceptBlockData {
+            requester: valid_bill_identity_block_data(),
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_req_to_accept_block_data() {
+        let accept = valid_req_to_accept_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_requester(BillRequestToAcceptBlockData { requester: invalid_bill_identity_block_data(), ..valid_req_to_accept_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_signing_address(BillRequestToAcceptBlockData { signing_address: invalid_address(), ..valid_req_to_accept_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillRequestToAcceptBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_req_to_accept_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_req_to_accept_block_data(
+        #[case] block: BillRequestToAcceptBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_accept_block_data() -> BillAcceptBlockData {
+        BillAcceptBlockData {
+            accepter: valid_bill_identity_block_data(),
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_accept_block_data() {
+        let accept = valid_accept_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_accepter(BillAcceptBlockData { accepter: invalid_bill_identity_block_data(), ..valid_accept_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_signing_address(BillAcceptBlockData { signing_address: invalid_address(), ..valid_accept_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillAcceptBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_accept_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_accept_block_data(
+        #[case] block: BillAcceptBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_req_to_pay_block_data() -> BillRequestToPayBlockData {
+        BillRequestToPayBlockData {
+            requester: valid_bill_identity_block_data(),
+            currency: "sat".into(),
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_req_to_pay_block_data() {
+        let accept = valid_req_to_pay_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_requester(BillRequestToPayBlockData { requester: invalid_bill_identity_block_data(), ..valid_req_to_pay_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_currency(BillRequestToPayBlockData { currency: "invalidcurrency".into(), ..valid_req_to_pay_block_data() }, ValidationError::InvalidCurrency)]
+    #[case::invalid_signing_address(BillRequestToPayBlockData { signing_address: invalid_address(), ..valid_req_to_pay_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillRequestToPayBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_req_to_pay_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_req_to_pay_block_data(
+        #[case] block: BillRequestToPayBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_mint_block_data() -> BillMintBlockData {
+        BillMintBlockData {
+            endorser: valid_bill_identity_block_data(),
+            endorsee: valid_bill_identity_block_data(),
+            currency: "sat".into(),
+            sum: 500,
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_mint_block_data() {
+        let accept = valid_mint_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_endorser(BillMintBlockData { endorser: invalid_bill_identity_block_data(), ..valid_mint_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_endorsee(BillMintBlockData { endorser: invalid_bill_identity_block_data(), ..valid_mint_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_currency(BillMintBlockData { currency: "invalidcurrency".into(), ..valid_mint_block_data() }, ValidationError::InvalidCurrency)]
+    #[case::invalid_sum(BillMintBlockData { sum: 0, ..valid_mint_block_data() }, ValidationError::InvalidSum)]
+    #[case::invalid_signing_address(BillMintBlockData { signing_address: invalid_address(), ..valid_mint_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillMintBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_mint_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_mint_block_data(
+        #[case] block: BillMintBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_offer_to_sell_block_data() -> BillOfferToSellBlockData {
+        BillOfferToSellBlockData {
+            seller: valid_bill_identity_block_data(),
+            buyer: valid_bill_identity_block_data(),
+            currency: "sat".into(),
+            sum: 500,
+            payment_address: VALID_PAYMENT_ADDRESS_TESTNET.into(),
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_offer_to_sell_block_data() {
+        let accept = valid_offer_to_sell_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_seller(BillOfferToSellBlockData { seller: invalid_bill_identity_block_data(), ..valid_offer_to_sell_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_buyer(BillOfferToSellBlockData { buyer: invalid_bill_identity_block_data(), ..valid_offer_to_sell_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_currency(BillOfferToSellBlockData { currency: "invalidcurrency".into(), ..valid_offer_to_sell_block_data() }, ValidationError::InvalidCurrency)]
+    #[case::invalid_sum(BillOfferToSellBlockData { sum: 0, ..valid_offer_to_sell_block_data() }, ValidationError::InvalidSum)]
+    #[case::invalid_payment_address(BillOfferToSellBlockData { payment_address: "invalidaddress".into(), ..valid_offer_to_sell_block_data() }, ValidationError::InvalidPaymentAddress)]
+    #[case::invalid_signing_address(BillOfferToSellBlockData { signing_address: invalid_address(), ..valid_offer_to_sell_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillOfferToSellBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_offer_to_sell_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_offer_to_sell_block_data(
+        #[case] block: BillOfferToSellBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_sell_block_data() -> BillSellBlockData {
+        BillSellBlockData {
+            seller: valid_bill_identity_block_data(),
+            buyer: valid_bill_identity_block_data(),
+            currency: "sat".into(),
+            sum: 500,
+            payment_address: VALID_PAYMENT_ADDRESS_TESTNET.into(),
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_sell_block_data() {
+        let accept = valid_sell_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_seller(BillSellBlockData { seller: invalid_bill_identity_block_data(), ..valid_sell_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_buyer(BillSellBlockData { buyer: invalid_bill_identity_block_data(), ..valid_sell_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_currency(BillSellBlockData { currency: "invalidcurrency".into(), ..valid_sell_block_data() }, ValidationError::InvalidCurrency)]
+    #[case::invalid_sum(BillSellBlockData { sum: 0, ..valid_sell_block_data() }, ValidationError::InvalidSum)]
+    #[case::invalid_payment_address(BillSellBlockData { payment_address: "invalidaddress".into(), ..valid_sell_block_data() }, ValidationError::InvalidPaymentAddress)]
+    #[case::invalid_signing_address(BillSellBlockData { signing_address: invalid_address(), ..valid_sell_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillSellBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_sell_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_sell_block_data(
+        #[case] block: BillSellBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_endorse_block_data() -> BillEndorseBlockData {
+        BillEndorseBlockData {
+            endorser: valid_bill_identity_block_data(),
+            endorsee: valid_bill_identity_block_data(),
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_endorse_block_data() {
+        let accept = valid_endorse_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_endorser(BillEndorseBlockData { endorser: invalid_bill_identity_block_data(), ..valid_endorse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_endorsee(BillEndorseBlockData { endorsee: invalid_bill_identity_block_data(), ..valid_endorse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_signing_address(BillEndorseBlockData { signing_address: invalid_address(), ..valid_endorse_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillEndorseBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_endorse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_endorse_block_data(
+        #[case] block: BillEndorseBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_req_to_recourse_block_data() -> BillRequestRecourseBlockData {
+        BillRequestRecourseBlockData {
+            recourser: valid_bill_identity_block_data(),
+            recoursee: valid_bill_identity_block_data(),
+            currency: "sat".into(),
+            sum: 500,
+            recourse_reason: BillRecourseReasonBlockData::Pay,
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_req_to_recourse_block_data() {
+        let accept = valid_req_to_recourse_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_recourser(BillRequestRecourseBlockData { recourser: invalid_bill_identity_block_data(), ..valid_req_to_recourse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_recoursee(BillRequestRecourseBlockData { recoursee: invalid_bill_identity_block_data(), ..valid_req_to_recourse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_sum(BillRequestRecourseBlockData { sum: 0, ..valid_req_to_recourse_block_data() }, ValidationError::InvalidSum)]
+    #[case::invalid_payment_address(BillRequestRecourseBlockData { currency: "invalidcurrency".into(), ..valid_req_to_recourse_block_data() }, ValidationError::InvalidCurrency)]
+    #[case::invalid_signing_address(BillRequestRecourseBlockData { signing_address: invalid_address(), ..valid_req_to_recourse_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillRequestRecourseBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_req_to_recourse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_req_to_recourse_block_data(
+        #[case] block: BillRequestRecourseBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_recourse_block_data() -> BillRecourseBlockData {
+        BillRecourseBlockData {
+            recourser: valid_bill_identity_block_data(),
+            recoursee: valid_bill_identity_block_data(),
+            currency: "sat".into(),
+            sum: 500,
+            recourse_reason: BillRecourseReasonBlockData::Pay,
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_recourse_block_data() {
+        let accept = valid_recourse_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_recourser(BillRecourseBlockData { recourser: invalid_bill_identity_block_data(), ..valid_recourse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_recoursee(BillRecourseBlockData { recoursee: invalid_bill_identity_block_data(), ..valid_recourse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_sum(BillRecourseBlockData { sum: 0, ..valid_recourse_block_data() }, ValidationError::InvalidSum)]
+    #[case::invalid_payment_address(BillRecourseBlockData { currency: "invalidcurrency".into(), ..valid_recourse_block_data() }, ValidationError::InvalidCurrency)]
+    #[case::invalid_signing_address(BillRecourseBlockData { signing_address: invalid_address(), ..valid_recourse_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillRecourseBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_recourse_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_recourse_block_data(
+        #[case] block: BillRecourseBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
+    }
+
+    fn valid_reject_block_data() -> BillRejectBlockData {
+        BillRejectBlockData {
+            rejecter: valid_bill_identity_block_data(),
+            signatory: Some(valid_bill_signatory_block_data()),
+            signing_timestamp: 1731593928,
+            signing_address: valid_address(),
+        }
+    }
+
+    #[test]
+    fn test_valid_reject_block_data() {
+        let accept = valid_reject_block_data();
+        assert_eq!(accept.validate(), Ok(()));
+    }
+
+    #[rstest]
+    #[case::invalid_rejecter(BillRejectBlockData { rejecter: invalid_bill_identity_block_data(), ..valid_reject_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    #[case::invalid_signing_address(BillRejectBlockData { signing_address: invalid_address(), ..valid_reject_block_data() }, ValidationError::FieldEmpty(Field::Country))]
+    #[case::invalid_signatory(BillRejectBlockData { signatory: Some(invalid_bill_signatory_block_data()), ..valid_reject_block_data() }, ValidationError::FieldEmpty(Field::Name))]
+    fn test_invalid_reject_block_data(
+        #[case] block: BillRejectBlockData,
+        #[case] expected_error: ValidationError,
+    ) {
+        assert_eq!(block.validate(), Err(expected_error));
     }
 }